Implements base physics manager for collision detection

+++ Minigin/structs.h
@@ -11,8 +11,8 @@ struct FVector2 {
 };
 
 struct Rect {
-	int x;
-	int y;
-	int w;
-	int h;
+	float x;
+	float y;
+	float w;
+	float h;
 };
